Tpetra::CrsMatrix: Attempt to fix issue in "classic" Gauss-Seidel.
The Tpetra specialization of Belos::MultiVecTraits sets all
Tpetra::MultiVector instances to have view semantics.  This means that
Tpetra::MultiVector::operator= always does a shallow copy when Tpetra
is used in Belos.  This works trivially fine with the new (Kokkos
refactor) version of Tpetra, which ONLY implements view semantics.
However, the implementation of Gauss-Seidel in the "classic" version
of Tpetra had a couple places that assumed operator= does a deep copy.
This is fine in isolation, but NOT fine when using Belos in an Ifpack2
test.

Tpetra::deep_copy works with both the classic and new versions of
Tpetra.  I changed the "classic" version of Gauss-Seidel to use this
when doing deep copies between MultiVector instances, rather than
using operator=.

NOTE: With the classic version of Tpetra, the Ifpack2 test for
multiple sweeps of Gauss-Seidel still fails!  It does NOT fail with
the new version of Tpetra.

I also fixed a latent issue in a possibly untested branch of the
implementation of Gauss-Seidel in the new version of Tpetra.