Tpetra::CrsMatrix: Work around Teuchos::ArrayView issue.
In a debug build, Teuchos::ArrayView's constructor that takes a raw
pointer and a length throws, if the length is zero and the row pointer
is not NULL.  While that may make sense for debugging, I have gotten a
report of this causing issues in some places in CrsMatrix.  This
commit works around the issue by passing in NULL explicitly to
Teuchos::ArrayView's constructor if the length is zero.