TpetraKernels,TpetraCore: Fixed KokkosBlas::nrm2_squared() interface.
It is now consistent with the intended KokkosBlas interface, in that
it is compile-time polymorphic on the input and output View ranks.
There are no longer two separately named interfaces for the
multivector and single-vector cases.

KokkosBlas::Impl::Nrm2_MV is called Nrm2_MV and not Nrm2Squared_MV,
because we might still want to use it to implement nrm2() (by calling
a "final" operation on the reduction result).  We have not implemented
this option yet.