Tpetra::CrsMatrix: Added a FIXME comment for Bug 6332.
This is a subtle unpackAndCombine() bug.  For each row of incoming
data, Tpetra::CrsMatrix::unpackAndCombineImpl() (called by
unpackAndCombine()) currently keeps track of what it claims to be the
maximum possible number of entries in that row.  This is only
nontrivial if the target matrix (*this) has a column Map.  If it does,
unpackAndCombineImpl() claims that the bound is
this->getColMap()->getNodeNumElements().

This upper bound is incorrect, because it does not account for the
"column Map filtering" feature of Import / Export on a CrsGraph or
CrsMatrix, when the target graph / matrix has a column Map.  "Column
Map filtering" means that the target matrix ignores incoming column
indices that are not in its column Map, if the target matrix actually
has a column Map.  Both Epetra and Tpetra have this feature.

A user manifested this bug.  Their target matrix has an overlapping
row Map M1 with the same Map M1 as the column Map, and their source
matrix has a nonoverlapping row Map and a conventional column Map.
The goal in that case is for the target matrix to filter out some
incoming entries, so that it end up "conceptually square."  The effect
is very much like that of Ifpack2::LocalFilter (which is buggy for a
likely different reason -- see Bug 5992).

This commit just adds a comment; it doesn't fix the bug.  My next
commit will fix the bug.