TrilinosCouplings:  Updates to FENL example.
Update the LocalViewTraits specialization for the Ensemble scalar type
on Cuda so that it is actually picked up.  Specializing on Cuda
execution space doesn't quite do it.  Instead I did the "Enabled = void"
pattern.  Also updated the print flag so that printing on Cuda is also
available (when using UVM).