Ifpack2::AdditiveSchwarz: Added more debug code to help fix failing test
The debug code shows that Y already has 2-norm Inf or NaN at the top
of iteration 0.  That's actually correct and to be expected if Y is
uninitialized and beta = 0.  The apply() method _should_ overwrite Y
in that case.  Perhaps it's not, though.  If not, that would be the
likely cause of the test failure.