Ifpack2::AdditiveSchwarz: Fixed failing AdditiveSchwarz_RILUK test.
See my previous commit's message.  I found that the output multivector
Y Y already had 2-norm Inf or NaN at the top of iteration 0 in
Ifpack2::AdditiveSchwarz::apply.  This should be fine if
ZeroStartingSolution_ is true, which it was in the failing test.  In
that case, the apply() method should have overwritten the contents of
Y before letting them "contaminate" other multivectors in the method.
Lines 605-8, which are commented out, do this.  The fact that they are
commented out made me a bit suspicious: it looks like the developer
who revised this method expected that it would be correct regardless
of the value of ZeroStartingSolution_.  I left them commented out, and
instead added a Y.putScalar(ZERO) if ZeroStartingSolution_ is true,
near the beginning of the method (outside of the iteration loop, lines
542-5).  This made the test pass.