Tpetra::Details::FixedHashTable: Now uses Kokkos not ArrayRCP inside.
FixedHashTable is what Tpetra::Map uses to convert a global index to a
local index, for a general ("noncontiguous") Map.  FixedHashTable has
a black-box interface, so it's easy to replace its insides.  The end
goal is to provide a "Kokkos refactor" version of Tpetra::Map that can
be used with Kokkos thread-parallel kernels.

In this commit, FixedHashTable always allocates host Views.  This is
because all computations and look-ups currently happen on the host.
While CUDA UVM lets host code access device memory, this access is
expensive compared with host memory access.  I don't want current CUDA
code that uses Tpetra to get slower.  Once we move FixedHashTable
setup and look-ups onto the device, it will make more sense to move
the Views into device memory.