Tpetra: Added memory_space typedef to KokkosDeviceWrapperNode.
KokkosClassic::SerialNode already has both execution_space and
memory_space typedefs, but Kokkos::Compat::KokkosDeviceWrapperNode
only has an execution_space typedef.  This commit does two things to
KokkosDeviceWrapperNode:

  1. Add a second template parameter, MemorySpace, which defaults to
     ExecutionSpace::memory_space
  2. Add a memory_space public typedef

The two template parameters of KokkosDeviceWrapperNode are the same
and occur in the same order as those of Kokkos::Device.  This is
deliberate.  The eventual goal in Tpetra is to replace "Node" with
"Device".  This commit helps that goal.