Tpetra::Details::FixedHashTable: Possible Bug 6335 fix.
The issue is that FixedHashTable's default execution space was Kokkos'
default execution space, NOT Tpetra's default execution space.  For
example, when Kokkos::Serial is Tpetra's default execution space,
OpenMP is off, and Kokkos::Threads is enabled, Kokkos::Threads is
Kokkos' default execution space.  However, Tpetra was never enabling
this space.  This was a problem for Kokkos::View initialization, which
was using Kokkos' default execution space for first-touch fill.

This commit removes the default value of the DeviceType template
parameter from FixedHashTable.  Now Tpetra must specify this type
explicitly.  This is OK because FixedHashTable is an implementation
detail of Tpetra.  It was only used in Tpetra::Map and in the
implementation of Tpetra::Directory.