Stokhos:  Add overload for memcpy() for PCE scalar type.
Tpetra uses memcpy() in one place as a workaround for gcc 4.9's default
alignment requirements, and this was causing trouble when using the PCE
scalar type.  A simple fix for this was to provide an overload of
memcpy() for this type.  Hopefully in the future the memcpy() will be
removed.