Sundance: Fixed build error (probably C++11 related).
The use case compared an std::ifstream to 0, as a test for validity.
This caused a build error with Clang 3.5 on Mac and C++11 enabled.  The
right way to do this is to invoke std::ifstream::operator!.  I did that
and the code built.  This code may have built with C++98 but it doesn't
anymore, so I changed it.