Tpetra: Fixed writeOperator test; uncovered writeOperator bug.
NOTE: One Tpetra test fails as a result of this commit.  It fails
because the test was not exercising a buggy case before.  Let this test
fail; I plan to fix it in my next commit.

Tpetra::MatrixMarket::Writer::writeOperator writes a Tpetra::Operator in
Matrix Market sparse format.  It has two overloads: one that writes to a
file, and one that writes to an std::ostream.

I discovered that the writeOperator test writes temporary files.  This
caused unwarranted test failures on my laptop.  It is unwise for tests
to write to files, so I rewrote the test to use std::ostringstream and
std::istringstream instead of file I/O.  When I did this, I changed the
test to invoke the overload of writeOperator that takes an std::ostream,
rather than the overload that takes a file name.  This made the test
fail!

I looked at the implementation of writeOperator, and noticed that only
the version that writes to a file actually writes the header!  The
version that writes to an output stream omits this; it does not write
the header.  It looks like the implementers did not realize that some
users might actually like to call the writeOperator(std::ostream&, ...)
method, even though it is public.

The right fix for this is to rename the current
writeOperator(std::ostream&, ...) method to writeOperatorImpl, and
provide a writeOperator(std::ostream&, ...) that looks like the current
writeOperator(const std::string& fileName, ...) overload.  I have not
done this yet; I will do it in my next commit.