Tpetra: Added comment noting CUDA issue in FixedHashTable.
Tpetra::Details::FixedHashTable has a "copy" constructor that accepts a
FixedHashTable with a different Device type.  This functor does a
Kokkos::deep_copy from the input's ptr_ to the output's ptr_.  The
problem is that ptr_ is a Kokkos::View of offset_type, and offset_type
depends on the memory space's size_type.  Kokkos::CudaSpace has a
differently sized size_type than Kokkos::HostSpace.  This means that
Kokkos::deep_copy won't build if CUDA is enabled.

This commit does not fix the issue, but it adds a comment to the
relevant line of code that explains _how_ to fix the issue.  I will fix
the issue in later commits.