Tpetra: FixedHashTable: Added option to ctors to keep keys.
I added an optional argument 'keepKeys" to some of the constructors of
Tpetra::Details::FixedHashTable.  If keepKeys = true, FixedHashTable
keeps (a deep copy of) the keys.  This will let FixedHashTable implement
the "reverse" map from values to keys.  Once FixedHashTable becomes the
"local Map," that will provide the equivalent of
Tpetra::Map::getGlobalElement.

The reverse map is harder to implement for the version of FixedHashTable
that takes arrays of both keys _and_ values.  However, we do not need
the reverse map in that case.  Thus, that constructor does not accept
the keepKeys option.