Tpetra: Added Details::OrdinalTraits traits class.
Tpetra makes extensive use of Teuchos::OrdinalTraits<T>::invalid() for T
= LocalOrdinal and GlobalOrdinal.  (It returns an "invalid" value which
Tpetra uses as a flag, both internally and for users).  This is NOT
marked as a Kokkos device function, and I'm not keen on fiddling with
Teuchos.  I can't use std::numeric_limits for the same reason.
Furthermore, it is confusing for users to have to reach into another
namespace when dealing with something as basic as flag values.

This commit adds the traits class Tpetra::Details::OrdinalTraits.  It
currently contains ONLY invalid().  For all T,
Tpetra::Details::OrdinalTraits<T>::invalid() ==
Teuchos::OrdinalTraits<T>::invalid().  This ensures backwards
compatibility.  Note that the class is NOT yet exposed to users; this is
why it currently lives in Tpetra::Details and not in Tpetra.

The reason I'm adding this class now is because I don't want
Tpetra::Details::FixedHashTable to need fields for storing flag values
of KeyType and ValueType.  I want FixedHashTable to be as lightweight as
possible, because it will become the "local Map" -- the object that goes
into Kokkos kernels and provides a subset of Tpetra::Map functionality.